# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: SEMLINK API
  version: 2.0.0
  description: REST contract for the SEMLINK semantic integration framework.
servers:
  - url: http://localhost:8080
paths:
...
  /api/v1/connections:
    get:
      summary: List registered database connections
      responses:
        "200":
          description: Connections
    post:
      summary: Register a database connection
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConnectionConfig"
      responses:
        "201":
          description: Connection registered
  /api/v1/connections/{id}:
    delete:
      summary: Remove a connection
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Removed
...
  /api/v1/query/sparql:
    post:
      summary: Execute SPARQL against the semantic layer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SparqlRequest"
      responses:
        "200":
          description: Query result
  /api/v1/query/natural:
    post:
      summary: Translate natural language to SPARQL and execute
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NaturalQueryRequest"
      responses:
        "200":
          description: Query result plus generated SPARQL
...
components:
  schemas:
    ConnectionConfig:
      type: object
      required: [id, type]
      properties:
        id:
          type: string
        type:
          type: string
          enum: [mysql, postgresql, oracle, sqlite, mongodb, neo4j, redis, cassandra, owl, csv]
        host:
          type: string
        port:
          type: integer
        database:
          type: string
        username:
          type: string
        ssl:
          type: boolean
        timeoutSeconds:
          type: integer
        options:
          type: object
          additionalProperties:
            type: string
    SparqlRequest:
      type: object
      required: [query]
      properties:
        query:
          type: string
        explain:
          type: boolean
    NaturalQueryRequest:
      type: object
      required: [question]
      properties:
        question:
          type: string
        showGeneratedSparql:
          type: boolean